requirements.txt
chardet==5.2.0
pandas==2.2.2
python-dotenv==1.0.1
git+https://github.com/agustin-mayer/livepanel-sdk